Заменено хранение пароля из userDefaults в keyChain
